requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.3
Pillow==10.1.0
reportlab==4.0.7
requests==2.31.0
firebase-admin>=7.0.0